Updated the build system for Xcode to be more versatile and reliable.
Xcode has traditionally used a Perl script, build-llvm.pl, to build LLDB on
behalf of the DebugClang, Debug, and Release configurations.  This script now
works when building with build-script (i.e., in the CustomSwift-* build
configurations) but DebugClang, Debug, and Release are broken.  These turn out
to be useful, and this patch fixes them.

One fix could have been to fix build-llvm.pl, but the problem extended beyond
this tool: package-clang-headers and package-swift-resources were also affected,
as was the logic for building the Swift REPL.  build-llvm.pl's design over time
had gone from clean and single-purpose to convoluted and brittle, partially due
to an imperative, loop-heavy architecture.

Instead, I have introduced lldbbuild.py, a Python library that implements many
common functions that Xcode builds can use to interface with the Swift build
system.  In addition to facilitating a drop-in replacement for build-llvm.pl
that re-enables the DebugClang, Debug, and Release configurations, lldbbuild.py
also acts as a facilitator for the other scripts that interface between Xcode
and the Swift build.

The intent is that you should not notice this change.  Aside from a bugfix in
prepare_binding_Python.py, this should not affect the Linux build at all.  Let
me know if you see any problems as a result of this switch.

apple-llvm-split-commit: 700f6c1143acb63ff4a42486e6bb4bd4b3d141d3
apple-llvm-split-dir: lldb/