Collect IRExecutionUnits as part of persistent expression state.
IRExecutionUnits contain code and data that persistent declarations can
depend on.  In order to keep them alive and provide for lookup of these
symbols, we now allow any PersistentExpressionState to keep a list of
execution units.  Then, when doing symbol lookup on behalf of an
expression, any IRExecutionUnit can consult the persistent expression
states on a particular Target to find the appropriate symbol.

<rdar://problem/22864976>

git-svn-id: https://llvm.org/svn/llvm-project/lldb/trunk@263995 91177308-0d34-0410-b5e6-96231b3b80d8
apple-llvm-split-commit: bbbcc910dc7da63f53ad8523389abef214765703
apple-llvm-split-dir: lldb/