temporary fix for static binding issue on non-Apple platforms
When swig static bindings are being used (i.e. when swig is not
present on the system), we use static bindings generated on
OS X.  We have introduced swig Python binding typemap code that
uses #ifdef __APPLE__ in the typemap definitions.  Those
entries look like they are interpreted at swig generation time,
not during compilation of the output from swig.  Thus, the static
files generated on OS X contain already-baked-in #ifdef
evaluation that is incorrect for Ubuntu and other non-Apple
platforms.

This is a temporary fix for:
https://bugs.swift.org/browse/SR-1192
(aka rdar://25626302)

apple-llvm-split-commit: fbbacbd780eb113a04fb99296cc1f6bb98c6d9b9
apple-llvm-split-dir: lldb/