SwiftASTContext: ensure modules have a resource dir set
With the fix that is about to go in for importing modules in
the REPL for Ubuntu, this exposes a case where module SwiftASTContexts
did not have a resource dir set.  This causes an assert to fire in
the Swift.

This change keeps track of the first resource dir set for a target,
and then uses that resource dir for modules if they don't otherwise
have a resource dir set.

This should be further reviewed by Sean Callanan, but shoul be
good enough to get our Ubuntu packages built and working once
I make the other related check-in into Swift that is dependent
on this one.

apple-llvm-split-commit: ea2a5f26e5c8f2aae576b2a2a6178d7415cb13b8
apple-llvm-split-dir: lldb/