Add the ability to limit "source regexp" breakpoints to a particular function within a source file.
This isn't done, I need to make the name match smarter (right now it requires an
exact match which is annoying for methods of a class in a namespace.

Also, though we use it in tests all over the place, it doesn't look like we have
a test for Source Regexp breakpoints by themselves, I'll add that in a follow-on patch.


git-svn-id: https://llvm.org/svn/llvm-project/lldb/trunk@267834 91177308-0d34-0410-b5e6-96231b3b80d8
apple-llvm-split-commit: 223633836a2d953823f33371c67426542ce61843
apple-llvm-split-dir: lldb/