[fix] Fixed a bug where const this would cause parser errors about $__lldb_expr.
In templated const functions, trying to run an expression would produce the
error

error: out-of-line definition of '$__lldb_expr' does not match any declaration
in 'foo' member declaration does not match because it is const qualified
error: 1 error parsing expression

which is no good.  It turned out we don't actually need to worry about "const,"
we just need to be consistent about the declaration of the expression and the
FunctionDecl we inject into the class for "this."

Also added a test case.

<rdar://problem/24985958>

git-svn-id: https://llvm.org/svn/llvm-project/lldb/trunk@268083 91177308-0d34-0410-b5e6-96231b3b80d8
apple-llvm-split-commit: 94dab0db200ddef22d208783498827e65d7e0548
apple-llvm-split-dir: lldb/