Revive the error message from "process load" and SBProcess::LoadImage.
IsPointedCString has problems with ValueObjects of type eTypeHostAddress.  We should
figure out the right thing to do in that case, but the test is silly here because we're
reading a type we've defined, so we know it is a const char *, and if the memory is good, 
we won't be able to read any characters, when we do ReadPointedString.

<rdar://problem/26612812>


git-svn-id: https://llvm.org/svn/llvm-project/lldb/trunk@272087 91177308-0d34-0410-b5e6-96231b3b80d8
apple-llvm-split-commit: da7077d43e35725c71dec31e776e25f658c79ce2
apple-llvm-split-dir: lldb/