prepend XFAIL/XPASS to the xUnit-reported test method name
If the xUnit formatter is instructed to mark XFAIL/XPASS results as
xUnit successes, then prepend XFAIL: or XPASS: to the method name.

Prior to this change and the corresponding change in
swift/utils/build-script-impl to mark XPASS and XFAIL results
as successes, the pretty-ified output from the Jenkins XUnit
plugin did not know anything about XFAILs or XPASSes.
With these two changes, we will see XPASS and XFAIL results
as "passing" in the xUnit results, but we'll be able to distinguish
them by the prefix.

rdar://26769429

apple-llvm-split-commit: 63785cc3d21db62420104563fd41095c65d8333f
apple-llvm-split-dir: lldb/