Warning about debugging optimized code was not happening without dSYMs. Now it works for DWARF in .o files on Darwin.
I changed "m_is_optimized" in lldb_private::CompileUnit over to be a lldb::LazyBool so that it can be set to eLazyBoolCalculate if it needs to be parsed later. With SymbolFileDWARFDebugMap, we don't actually open the DWARF in the .o files for each compile unit until later, and we can't tell if a compile unit is optimized ahead of time. So to avoid pulling in all .o right away just so we can answer the questions of "is this compile unit optimized" we defer it until a point where we will have the compile unit parsed.

<rdar://problem/26068360>

git-svn-id: https://llvm.org/svn/llvm-project/lldb/trunk@274585 91177308-0d34-0410-b5e6-96231b3b80d8
(cherry picked from commit 8ea617d64bc3a3020033f539588becc0057beddb)

apple-llvm-split-commit: eb50d13bb5e442b4771c8185e5504b6371409c1d
apple-llvm-split-dir: lldb/