Updated the way we figure out how to rewrite expressions.
Right now we have a very tangled relationship between ScanContext() in
[Clang,Swift]UserExpression and ExpressionSourceCode::GetText.  There are also
some tendrils in SwiftExpressionParser where types get adjusted.

TestWeakSelf.py has been failing for a long time, and fixing that would have
added yet another rickety hack on top of this structure, so I cleaned it up a
bit.  The only cross-cutting change is that instead of a bunch of flags common
to C, C++, and Swift, we now have language-specific flags.

<rdar://problem/22738516>

apple-llvm-split-commit: 5f3aa16203d308773108c4d03ed4d4ec8e4a47c2
apple-llvm-split-dir: lldb/