Don't deadlock if someone tries to run an expression while another is in flight.
This is a short-term fix for a problem where the ObjC runtime ISA fetching code
kicks in while we are still in RunThreadPlan for another expression.  This will
cause the second expression evaluation to fail right away.  We need a mechanism
to yield ability to run an expression from one actor to another, but that's for
another day.

<rdar://problem/27551912>

apple-llvm-split-commit: 8993c5c0b0b9f8198af32d0ce6e46a2b37718a97
apple-llvm-split-dir: lldb/