Ignore uninitialized generic variables in expressions, fixing a spurious error.
The Swift expression parser knows to ignore variables that can't be read.
However, variables of generic type, when uninitialized, appear to be readable
until you try to get their fully-resolved type.  This was causing the expression
parser to erroneously expose them to expressions even though they are unusable.
This patch fixes that and adds a testcase.

<rdar://problem/27669166>

apple-llvm-split-commit: 8d8bb4f0505f88951da9723bf14c6770e0f089d3
apple-llvm-split-dir: lldb/