[CMake] NFC. Cleanup repl_swift targets
This change does the following cleanup:

* Move linux-specific flags to a variable that is only set on linux
* Collapse the add_custom_command_target calls into one call
* Remove explicit add_dependencies calls in favor of using DEPENDS arguments
* Wrap COMMAND arguments so that it is easier to read
* Explicitly name the repl_swift target

Since CMake variables that are unset expand to nothing I generally encourage the use of variables that are conditionally set to avoid duplicating code wherever possible. In this instance using the linux_only_flags variable allows the add_custom_command_target calls to be collapsed, so modifications to the target or command definition don't need to be duplicated.

Also the add_dependencies calls are unnecessary if we use the DEPENDS arguments on the custom command and target calls. In CMake targets do not need to be created in order to setup dependencies, CMake will warn or error at the end of generation if it could not resolve a target dependency.

The last change here is providing an explicit name to the repl_swift target. When you do not specify an explicit target name Swift's add_custom_command_target gives you a generated target with an MD5 in it which is completely inscrutable. For debugging the build system and easy incremental builds of the repl_swift tool it is better to have a named target.

apple-llvm-split-commit: 3e84c1e4cb4bce24861c87ca8bc76d8ec6a3c6c5
apple-llvm-split-dir: lldb/