Fix arm64 sub-register definitions
The "value regs" field was filled incorrectly. It is supposed to list the
registers that *this* register is a sub-register of, not the other way around.
This manifested itself in "register read" showing only the smaller sub-registers
(and a bunch of tests not passing). I am not sure if the "invalidates" field is
correct either, but it's usage seems to be inconsistent, so I'll leave that as-is
for now.

git-svn-id: https://llvm.org/svn/llvm-project/lldb/trunk@284981 91177308-0d34-0410-b5e6-96231b3b80d8
(cherry picked from commit a1e6708fa6f1b47afe62e607ac944345388ee6b2)

apple-llvm-split-commit: 85e9e3cafb8cd2c58e1197510f48ac13e92fbde3
apple-llvm-split-dir: lldb/