[Test Suite] Properly respect --framework option
Summary:
dotest.py has a framework option that is not respected. This patch makes the framework path properly configurable via the --framework option.

This patch also adds a function to the lldbtest.Base class named "hasDarwinFramework" which allows us to not rely on the host platform to determine if a framework is present. If running on Darwin, and not building a framework, this will follow the *nix code paths which are appropriate for Darwin.

Reviewers: tfiala

Subscribers: lldb-commits, mgorny

Differential Revision: https://reviews.llvm.org/D25886

git-svn-id: https://llvm.org/svn/llvm-project/lldb/trunk@285541 91177308-0d34-0410-b5e6-96231b3b80d8
apple-llvm-split-commit: da606c55d4cbe377f9a8462f502971c7087c5f5e
apple-llvm-split-dir: lldb/