[CMake] Support distribution install for LLDB.framework
This patch adds the last bit of support to get LLVM_DISTRIBUTION_COMPONENTS working with libLLDB when built as a framework.

This patch adds dummy install targets for binaries built into the framework's Resources directory, and makes the framework's install target depend on all the binaries that get installed with the framework.

git-svn-id: https://llvm.org/svn/llvm-project/lldb/trunk@290273 91177308-0d34-0410-b5e6-96231b3b80d8
apple-llvm-split-commit: 1f877202edf45e09195ccf54d5539c5777f80481
apple-llvm-split-dir: lldb/