[dotest] Remove crashinfo hook
Summary:
This used to be important when all tests were run in a single process,
but that has no longer been the case for a while. Furthermore, this hook fails
to build on new mac versions for several people, and it's not clear
whether fixing it is worth the effort.

Reviewers: jingham, clayborg, davide

Subscribers: lldb-commits

Differential Revision: https://reviews.llvm.org/D41871

git-svn-id: https://llvm.org/svn/llvm-project/lldb/trunk@322167 91177308-0d34-0410-b5e6-96231b3b80d8
(cherry picked from commit 3a7b4021def1308e480995c3d053668b1e44408b)

apple-llvm-split-commit: 2ec5f0205edf74e296c13c22025869f3945e52ea
apple-llvm-split-dir: lldb/