Fix TestStopReplyContainsThreadPcs on 32-bit x86 (pr36013)
Summary:
The issue was that we were parsing the registers into 64-bit integers
and the calling swapByteOrder without regard for the actual size of the
register. This switches the test to use the RegisterValue class which
tracks the register size, and knows how to initialize itself from a
piece of memory (so we don't need to swap byte order ourselves).

Reviewers: eugene, davide

Subscribers: lldb-commits

Differential Revision: https://reviews.llvm.org/D43376

git-svn-id: https://llvm.org/svn/llvm-project/lldb/trunk@325511 91177308-0d34-0410-b5e6-96231b3b80d8
apple-llvm-split-commit: cd1ffbfda93d32c0af02ed222e1dd22a41b51c79
apple-llvm-split-dir: lldb/