[Dictionary] Rewrite the test added in r327587 as an inline test.
Until we have a better story for putting commands and check lines
in the same file (they're currently ignored), it seems that inline
tests are actually more concise and easier to understand.
Too bad we have still some python boilerplate, but that's not
really substantial so we can live with it.

Thanks to Fred for pointing out and Jim for explaining me how
to use the inline test format.

<rdar://problem/34806516>

git-svn-id: https://llvm.org/svn/llvm-project/lldb/trunk@327592 91177308-0d34-0410-b5e6-96231b3b80d8
apple-llvm-split-commit: dd00a9c94f3e0fc2d1ef71e519adbe3af508e4b7
apple-llvm-split-dir: lldb/