Swift: Add a testcases where the expression evaluator fails because of search path mismatches.
This is step 0 towards fixing these kinds of issues.

This test builds a .dylib whose .swiftmodule imports a Clang module
which uses a header file "Foo.h". The main swift program also imports
a Clang module that imports a header file called "Foo.h", but it is in
a different directory. We are using a headermap to switch between the
two versions of "Foo.h" during build time.

rdar://problem/38686915

apple-llvm-split-commit: de937fe4d40afdb87368cf04d95856a69252faab
apple-llvm-split-dir: lldb/