[Commands] Add a (currently empty) `stats` command.
This one will be used to print statistics about lldb sessions
(including, e.g. number of expression evaluation succeeded or
failed). I decided to commit the skeleton first so that we have
a clean reference on how a command should be implemented.
My future commits are going to populate this command and test
it.

<rdar://problem/36555975>

git-svn-id: https://llvm.org/svn/llvm-project/lldb/trunk@328378 91177308-0d34-0410-b5e6-96231b3b80d8
apple-llvm-split-commit: 5f7bf25e0eb1a5e6ee1899cc3bae3cfa35be5622
apple-llvm-split-dir: lldb/