Re-land "Don't assume backing thread shares protocol ID."
When we're dealing with virtual (memory) threads created by the OS
plugins, there's no guarantee that the real thread and the backing
thread share a protocol ID. Instead, we should iterate over the memory
threads to find the virtual thread that is backed by the current real
thread.

Differential revision: https://reviews.llvm.org/D45497

rdar://36485830

The original revision (r329891) was reverted because the associated
tests ran into a deadlock on the Linux bots. That problem was resolved
by r330002.

git-svn-id: https://llvm.org/svn/llvm-project/lldb/trunk@330005 91177308-0d34-0410-b5e6-96231b3b80d8
apple-llvm-split-commit: ccd3bb3dd9bdc80c89b75b441645738b4df5c018
apple-llvm-split-dir: lldb/