Backport the statistics command to swift.  (#531)
* [Commands] Add a (currently empty) `stats` command.

This one will be used to print statistics about lldb sessions
(including, e.g. number of expression evaluation succeeded or
failed). I decided to commit the skeleton first so that we have
a clean reference on how a command should be implemented.
My future commits are going to populate this command and test
it.

<rdar://problem/36555975>

git-svn-id: https://llvm.org/svn/llvm-project/lldb/trunk@328378 91177308-0d34-0410-b5e6-96231b3b80d8

* [Command] Implement `statistics` command.

This allows us to collect useful metrics about lldb debugging sessions.

I thought that an example would be better than a thousand words:

  Process 19705 stopped
  * thread #1, queue = 'com.apple.main-thread', stop reason = step in
      frame #0: 0x0000000100000fb4 blah`main at blah.c:3
     1    int main(void) {
     2      int a = 6;
  -> 3      return 0;
     4    }
  (lldb) statistics enable
  (lldb) frame var a
  (int) a = 6
  (lldb) expr a
  (int) $1 = 6
  (lldb) statistics disable
  (lldb) statistics dump
  Number of expr evaluation successes : 1
  Number of expr evaluation failures : 0
  Number of frame var successes : 1
  Number of frame var failures : 0

Future improvements might include:

1. Passing a file, or implementing categories. The way this patch has
been implemented is generic enough to allow this to be extended
easily without breaking the grammar.
2. Adding an SBAPI and Python API for use in scripts.

Thanks to Jim Ingham for discussing the design with me.

<rdar://problem/36555975>

Differential Revision:  https://reviews.llvm.org/D45547

git-svn-id: https://llvm.org/svn/llvm-project/lldb/trunk@330043 91177308-0d34-0410-b5e6-96231b3b80d8

* [Command] Simplify the code and make it less error prone. NFCI.

Pointed out by Jim.

git-svn-id: https://llvm.org/svn/llvm-project/lldb/trunk@330047 91177308-0d34-0410-b5e6-96231b3b80d8

* [Statistics] Move GetStatDescription to an header file.

In preparation for using it in the API.

<rdar://problem/36555975>

git-svn-id: https://llvm.org/svn/llvm-project/lldb/trunk@330147 91177308-0d34-0410-b5e6-96231b3b80d8

* [lldb-private] Mark a function as inline to prevent multiple def errors.

git-svn-id: https://llvm.org/svn/llvm-project/lldb/trunk@330149 91177308-0d34-0410-b5e6-96231b3b80d8

* [Commands] Expose statistics through the SBAPI.

The API is `SBStructuredData GetStatistics()`. This allows
the command to be used in scripts.

<rdar://problem/36555975>

git-svn-id: https://llvm.org/svn/llvm-project/lldb/trunk@330165 91177308-0d34-0410-b5e6-96231b3b80d8

* [Statistics] Relax the test output a little bit.

We don't really care about the order as this is a dictionary.
It should be more resilient to changes (adding/shuffling stats
around).

Pointed out by Jason Molenda in a post-commit review (thanks Jason).

git-svn-id: https://llvm.org/svn/llvm-project/lldb/trunk@330170 91177308-0d34-0410-b5e6-96231b3b80d8

* Fix an xcodeproject mismerge.

apple-llvm-split-commit: f9d401a36a09b68b632d576ad0f3b7eb3dce388f
apple-llvm-split-dir: lldb/