Change the default for target.use-all-compiler-flags to true.
This is a safe thing to do now since we have a reliable fallback
mechanism in the module-specific scract AST contexts. The main
difference for the end user that are affected by this change will be
that they don't get a confusing error message befopre the first
expression is evaluated.

<rdar://problem/39243524>

apple-llvm-split-commit: 45083652a628c66efa6b4365ec305b682d82f7a6
apple-llvm-split-dir: lldb/