[CMake] Set the default test compiler deterministically
For swift-lldb, switch the default testing behavior so we use the
in-tree compilers by default.

Also, stop setting the cached LLDB_TEST_C_COMPILER variable multiple
times. Just set it one time to the compiler we want to test with.

This address a non-deterministic issue which sometimes caused
LLDB_TEST_C_COMPILER to be set to the empty string. My theory for how
this happened is:

We set the cached variable LLDB_TEST_C_COMPILER for the first time
in the top-level cmake file, and then

We set the variable again in lit/CMakeLists.txt.

The second set did not use the FORCE modifier, so the empty string
should never have been assigned to LLDB_TEST_C_COMPILER, but perhaps
there was a bug in the version of cmake installed on the Linux bots
where we sporadically saw this issue.

At any rate, this PR should address the non-determinism by removing the
multiple assignments to the cached variable.

rdar://39537655, rdar://38935299

(cherry picked from commit ac3f72dc7a6fa3de46530b45cc74631d26c448bd)

 Conflicts:
	CMakeLists.txt

apple-llvm-split-commit: 99ac9ef97e8d830e4d04af77e77e522760acc976
apple-llvm-split-dir: lldb/