[DataFmt] Fix formatting of bridged CFStrings
The formatter for Swift strings called CreateValueObjectFromAddress in
order to construct a ValueObject for bridged strings. In an unknown set
of cases, this API call can accidentally dereference the payload, giving
an invalid ValueObject that the NSString summary provider doesn't
understand.

Address the issue by making sure the address is never dereferenced when
constructing the ValueObject for bridged strings. As a follow-up, there
is a radar to investigate why the accidental dereference occurs
(r://39741576)

rdar://39710387

apple-llvm-split-commit: f945d628ca1119a83561d7755d6d65393a85f5a1
apple-llvm-split-dir: lldb/