[debugserver] Fix handling of the 'g' packet
LLDB doesn't use this packet so we never hit this, but it looks like
some other projects talk to debugserver and are hitting an assert
(https://github.com/derekparker/delve/issues/1015).

We had an off by 1 in the accounting of the FPU structure sizes.
I added a test that basically just check that 'g' doesn't return
an error (currently it assert in debug builds). I didn't make
it an lldb-server test because it looks like lldb-server doesn't
implement the g packet.

git-svn-id: https://llvm.org/svn/llvm-project/lldb/trunk@331004 91177308-0d34-0410-b5e6-96231b3b80d8
apple-llvm-split-commit: b3750e58a6dd9705d87dcb6c2d16c448f9ab054d
apple-llvm-split-dir: lldb/