[CMake] Move swift compiler logic
Instead of duplicating the logic that passes the swift compiler and
library path, move it into test/CMakeLists.txt and pass it as a regular
argument to lldb-dotest.py.

This also gets rids of the remaining generator expressions which are
incompatible with the way we currently generate the lldb-dotest wrapper
and lit runner for the test suite.

apple-llvm-split-commit: daa2eb72666e78be76b225e9a27e1a9a2e6b5f22
apple-llvm-split-dir: lldb/