[testsuite] Implement a category to skip libstdcxx tests
On systems where it's not supported.
As far as I understand Linux is the only systems which now ships
with libstdcxx (maybe NetBSD?, but I'm not entirely sure of the
state of lldb on the platform).
We could make this more fine grained looking for the header as
we do for libcxx. This is a little tricky as there's no such
thing as /usr/include/c++/v1, but libstdcxx encodes the version
number in the path (i.e. /usr/include/c++/5.4). I guess we might
match a regex, but it seems fragile to me.

Differential Revision:  https://reviews.llvm.org/D49110

git-svn-id: https://llvm.org/svn/llvm-project/lldb/trunk@336724 91177308-0d34-0410-b5e6-96231b3b80d8
apple-llvm-split-commit: 24dab60946c5d0b1effa4c56a01f06f2261ffd8b
apple-llvm-split-dir: lldb/