Prevent the Swift scratch AST context to be reset while it is being used.
This patch introduces a Reader-Writer-Lock around the scratch
context. Users of a scratch context acquire a read lock while they are
depending on a scratch context. Only if no readers a around is Target
allowed to reset a Swift scratch context and delete the old one.

The Swift scratch context may need to be replaced when it gets corrupted,
for example due to incompatible ClangImporter options. This locking
mechanism guarantees that this won't happen while a client is using the
context.

In Swift there are three use-cases for ASTContexts with
different requirements and guarantees.

- Module ASTContexts are used for the static type system. They
  are created once for each lldb::Module and live forever.

- Scratch AST Contexts are used for expressions
  (thus far everything is like in the Clang language support)

- Scratch AST Contexts are also used to express the results of
   any dynamic type resolution done by RemoteAST or Archetype
   binding.

Because expressions and dynamic type resolution may trigger the
import of another module, the scratch context may become
unusable. When a scratch context is in a fatal error state,
GetScratchSwiftASTContext() will create a fresh global context,
or even separate scratch contexts for each lldb::Module. But it
will only do this if no client holds on to a read lock on \b
m_scratch_typesystem_lock.

rdar://problem/42399818

apple-llvm-split-commit: 8d2c620ea07618284b8da074ba49a22e2a7becda
apple-llvm-split-dir: lldb/