Support setting a breakpoint by FileSpec+Line+Column in the SBAPI.
This patch extends the SBAPI to allow for setting a breakpoint not
only at a specific line, but also at a specific (minimum) column. When
a column is specified, it will try to find an exact match or the
closest match on the same line that comes after the specified
location.

Differential Revision: https://reviews.llvm.org/D51461

git-svn-id: https://llvm.org/svn/llvm-project/lldb/trunk@341078 91177308-0d34-0410-b5e6-96231b3b80d8
(cherry picked from commit 9b6b462a3d2f916531b0bf334879aa64079d5c01)

apple-llvm-split-commit: 686dcc4dd53eab7a1e242791a52288f86cf5e58e
apple-llvm-split-dir: lldb/