[NFC] Turn "load dependent files" boolean into an enum
This is an NFC commit to refactor the "load dependent files" parameter
from a boolean to an enum value. We want to be able to specify a
default, in which case we decide whether or not to load the dependent
files based on whether the target is an executable or not (i.e. a
dylib).

This is a dependency for D51934.

Differential revision: https://reviews.llvm.org/D51859

git-svn-id: https://llvm.org/svn/llvm-project/lldb/trunk@342633 91177308-0d34-0410-b5e6-96231b3b80d8
(cherry picked from commit cae2d34a12503d1d3c23172f00cf20bef63641fd)

apple-llvm-split-commit: 5612a9eeec0ab823586affed48ad69d6ca0729db
apple-llvm-split-dir: lldb/