Fix the layout calculation of generic tuple types.
Performing dynamic type resolution on a tuple type can actually change
the tuple's layout, so when computing the offsets of a tuple's
elements, we need to use the static (but archetype-bound) type instead
of the dynamic type.

This patch requires a matching commit in the Swift compiler.

rdar://problem/45462765

apple-llvm-split-commit: 852a86dba1663522d5ace5d441d4a3d530743070
apple-llvm-split-dir: lldb/