[Target] AutoImport modules when getting the scratch ASTContext.
This is easy, after a fair amount of yak shaving. The patch
fixes the case where we need to print something from foundation,
but we didn't import foundation already (i.e. this is the first `expr`).
Don't imprort if the ASTContext was already poisoned, it would be
a disaster.

Finally fixes <rdar://problem/30398933>

apple-llvm-split-commit: 17db68da6663b9203d31e5a226912a2f1b4df425
apple-llvm-split-dir: lldb/