[Swift] Remove GetDynamicTypeAndAddress_Tuple()
It's not really correct to unwrap existentials in tuple elements because
the resulting concrete tuple type has a different layout than the original
type.

This changes some test output but note that the old behavior was
inconsistent; we would only call GetDynamicTypeAndAddress_Tuple() if the
original type was already generic, so a type like (Error, T) for a generic
parameter 'T' would unwrap the Error payload, whereas a type like
(Error, Int) would not.

apple-llvm-split-commit: 6236b442908c559f52f3d9417a90b3e4a05fd45e
apple-llvm-split-dir: lldb/